backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
anthropic==0.5.0
python-multipart==0.0.6
httpx==0.25.0 